Conver apartments to SQLAlchemy
This is WRONG.  Setting for now so that it works, but moving
to DynamoDB in the future.
Note to self: Moving on so that I can get the rest of the project
working correctly so I can upload it to prod and use the task part of it.

This part needs to be re-done and added as a milestone.