refactor: Remove `validate_response` helper function
This function had the wrong abstraction.
It simply tested the status code and then assembled an error message
if the status code was not as expected, but also logged
the error message and flashed it using the Flask `flash` function.
This caused testing issues as well as not knowing what the function
would actually return, as either a generic message would need to be
returned for successful, or a message would need to be added after the
function anyway.

It was more clear to eliminate the function and specify the error
handling explicitely after each response.