feat(auth): move QueryAPI inside routes to catch current_user
Once the routes require a login, they will need to have the
current_user available, which was not available when QueryAPI was
instantiated outside of the request context.

Also updated the box_packing routes to use the new QueryAPI and
simplified a lot of the routes.  Combined index and box routes into
one route that can accept a box_id or not.

Updated the journal routes to use the new QueryAPI and added schemas.